Introduces heartbeats.
A heartbeat is sent periodically, i.e., every 5 minutes to all connected nodes. When a node receives a heartbeat, it must store it in its local leafSet. The leafSet must hold key-value pairs in the form of nodeID -> socketIdentifier. A socket identifier may be a referencey to a connected socket itself, or some way you may derive the socket.